pushed useEffect to today page
